ml-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
opencv-python==4.8.1.78
ultralytics==8.0.196
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2